r/test: moved retry policy definition out of fixture
Moved the definition of retry policy out of raft fixture not to depend
on cluster error code definitions in Raft submodule
