+++ src/v/cluster/tests/BUILD
@@ -39,3 +39,44 @@ redpanda_cc_btest(
         "@seastar//:testing",
     ],
 )
+
+redpanda_test_cc_library(
+    name = "raft_fixiture_retry_policy",
+    hdrs = [
+        "raft_fixture_retry_policy.h",
+    ],
+    include_prefix = "cluster/tests",
+    deps = [
+        "//src/v/cluster",
+        "//src/v/raft",
+    ],
+)
+
+redpanda_cc_gtest(
+    name = "partition_properties_stm_test",
+    timeout = "short",
+    srcs = [
+        "partition_properties_stm_test.cc",
+    ],
+    deps = [
+        ":raft_fixiture_retry_policy",
+        ":utils",
+        "//src/v/bytes:iobuf",
+        "//src/v/bytes:random",
+        "//src/v/cluster",
+        "//src/v/config",
+        "//src/v/model",
+        "//src/v/model/tests:random",
+        "//src/v/raft",
+        "//src/v/raft/tests:raft_fixture",
+        "//src/v/random:generators",
+        "//src/v/serde",
+        "//src/v/test_utils:gtest",
+        "//src/v/test_utils:seastar_boost",
+        "@abseil-cpp//absl/container:flat_hash_map",
+        "@abseil-cpp//absl/container:flat_hash_set",
+        "@googletest//:gtest",
+        "@seastar",
+        "@seastar//:testing",
+    ],
+)
