+++ src/v/cluster/tests/tm_stm_tests.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "cluster/tests/raft_fixture_retry_policy.h"
 #include "cluster/tm_stm.h"
 #include "kafka/protocol/types.h"
 #include "model/fundamental.h"
