+++ src/v/raft/tests/persisted_stm_test.cc
@@ -16,6 +16,7 @@
 #include "raft/group_configuration.h"
 #include "raft/persisted_stm.h"
 #include "raft/tests/raft_fixture.h"
+#include "raft/tests/raft_fixture_retry_policy.h"
 #include "raft/tests/stm_test_fixture.h"
 #include "raft/types.h"
 #include "random/generators.h"
